tools/c7n_tencent/pyproject.toml
[tool.poetry]
name = "c7n_tencent"
version = "0.1.0"
description = "Cloud Custodian - TENCENT Cloud Provider"
readme = "readme.md"
homepage = "https://cloudcustodian.io"
repository = "https://github.com/cloud-custodian/cloud-custodian"
documentation = "https://cloudcustodian.io/docs/"
authors = ["Cloud Custodian Project"]
license = "Apache-2.0"
classifiers = [
   "Topic :: System :: Systems Administration",
   "Topic :: System :: Distributed Computing"
]

[tool.poetry.dependencies]
python = "^3.7"
tencentcloud-sdk-python = "^3.0.234"
crcmod = "^1.7"
pycryptodome = "^3.9.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
